+++ src/Emulator/Cores/RiscV/PlatformLevelInterruptController.cs
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2010-2022 Antmicro
+// Copyright (c) 2010-2023 Antmicro
 //
 // This file is licensed under the MIT License.
 // Full license text is available in 'licenses/MIT.txt'.
@@ -50,7 +50,7 @@ namespace Antmicro.Renode.Peripherals.IRQControllers
                                     {
                                         if(prioritiesEnabled)
                                         {
-                                            irqSources[j].Priority = value;
+                                            irqSources[j].Priority = (uint)value;
                                             RefreshInterrupts();
                                         }
                                     });
