+++ src/Emulator/Cores/X86/LAPIC.cs
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2010-2018 Antmicro
+// Copyright (c) 2010-2023 Antmicro
 //
 // This file is licensed under the MIT License.
 // Full license text is available in 'licenses/MIT.txt'.
@@ -286,7 +286,7 @@ namespace Antmicro.Renode.Peripherals.IRQControllers
             }
         }
 
-        public uint InternalTimerVector { get => localTimerVector.Value; }
+        public uint InternalTimerVector { get => (uint)localTimerVector.Value; }
 
         private int FindPendingInterrupt()
         {
